Fix photo bug
Problem: Open a todo item with notes. Tap on camera to take a photo.
Now tap on camera again but hit cancel instead. When you go back to the
add to do item view, the original photo disappears

Fix:
Saved the photo into the model on navigate back from camera.